+++ Assets/Exanite.Arpg/Networking/Client/ConnectResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Exanite.Arpg.NewNetworking.Client
+﻿namespace Exanite.Arpg.Networking.Client
 {
     public class ConnectResult
     {
