+++ Assets/Exanite.Arpg/Networking/Client/ConnectedEventArgs.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using LiteNetLib;
 
-namespace Exanite.Arpg.NewNetworking.Client
+namespace Exanite.Arpg.Networking.Client
 {
     public class ConnectedEventArgs : EventArgs
     {
