+++ Assets/Exanite.Arpg/Networking/EventHandler.cs
@@ -1,4 +1,4 @@
-﻿namespace Exanite.Arpg.NewNetworking
+﻿namespace Exanite.Arpg.Networking
 {
     public delegate void EventHandler<TSender, TEventArgs>(TSender sender, TEventArgs e);
 }
