+++ Assets/Exanite.Arpg/Networking/IPacket.cs
@@ -1,6 +1,6 @@
 ﻿using LiteNetLib.Utils;
 
-namespace Exanite.Arpg.NewNetworking
+namespace Exanite.Arpg.Networking
 {
     public interface IPacket : INetSerializable { }
 }
