+++ Assets/Exanite.Arpg/Networking/NetSerializationExtensions.cs.meta
