+++ Assets/Prototype/LiteNetLib/Client/ClientGameManager.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using Exanite.Arpg.Logging;
-using Exanite.Arpg.NewNetworking.Client;
+using Exanite.Arpg.Networking.Client;
 using LiteNetLib;
 using Prototype.LiteNetLib.Shared;
 using Prototype.LiteNetLib.Shared.Packets;
