+++ Assets/Prototype/LiteNetLib/Client/PlayerController.cs
@@ -1,5 +1,5 @@
 ﻿
-using Exanite.Arpg.NewNetworking.Client;
+using Exanite.Arpg.Networking.Client;
 using LiteNetLib;
 using Prototype.LiteNetLib.Shared;
 using Prototype.LiteNetLib.Shared.Packets;
