+++ Assets/Prototype/LiteNetLib/Shared/Packets/PlayerInputPacket.cs
@@ -1,4 +1,4 @@
-﻿using Exanite.Arpg.NewNetworking;
+﻿using Exanite.Arpg.Networking;
 using LiteNetLib.Utils;
 using UnityEngine;
